backend/services/retrieval_service/requirements.txt
fastapi
uvicorn
pymongo
langchain-huggingface
sentence-transformers
python-dotenv
